backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
psycopg2-binary
dynaconf
alembic
requests
flask-migrate
python-dotenv
gunicorn
email-validator
validators
httpx
